LPS-56065 migrate form navigators
