+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/servlet/taglib/ui/CompanySettingsRatingsFormNavigatorEntry.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.portalsettings.web.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
 import com.liferay.portal.model.Company;
 import com.liferay.portal.model.User;
 import com.liferay.portlet.ratings.definition.PortletRatingsDefinitionUtil;
@@ -23,10 +23,19 @@ import com.liferay.portlet.ratings.definition.PortletRatingsDefinitionValues;
 
 import java.util.Map;
 
+import javax.servlet.ServletContext;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
 /**
  * @author Pei-Jung Lan
+ * @author Philip Jones
  */
-@OSGiBeanProperties(property = {"service.ranking:Integer=10"})
+@Component(
+		immediate = true, property = {"service.ranking:Integer=10"},
+		service = FormNavigatorEntry.class
+)
 public class CompanySettingsRatingsFormNavigatorEntry
 	extends BaseCompanySettingsFormNavigatorEntry {
 
@@ -56,8 +65,17 @@ public class CompanySettingsRatingsFormNavigatorEntry
 	}
 
 	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.portal.settings.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
+	}
+
+	@Override
 	protected String getJspPath() {
-		return "/html/portlet/portal_settings/ratings.jsp";
+		return "/ratings.jsp";
 	}
 
 }
