+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/servlet/taglib/ui/CompanySettingsRecycleBinFormNavigatorEntry.java
@@ -1,0 +1,68 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portalsettings.web.servlet.taglib.ui;
+
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
+import com.liferay.portal.model.Company;
+import com.liferay.portal.model.User;
+
+import javax.servlet.ServletContext;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Pei-Jung Lan
+ * @author Philip Jones
+ */
+@Component(
+		immediate = true, property = {"service.ranking:Integer=10"},
+		service = FormNavigatorEntry.class
+)
+public class CompanySettingsRecycleBinFormNavigatorEntry
+	extends BaseCompanySettingsFormNavigatorEntry {
+
+	@Override
+	public String getCategoryKey() {
+		return
+			FormNavigatorConstants.CATEGORY_KEY_COMPANY_SETTINGS_CONFIGURATION;
+	}
+
+	@Override
+	public String getKey() {
+		return "recycle-bin";
+	}
+
+	@Override
+	public boolean isVisible(User user, Company company) {
+		return false;
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.portal.settings.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
+	}
+
+	@Override
+	protected String getJspPath() {
+		return "/recycle_bin.jsp";
+	}
+
+}
