+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/servlet/taglib/ui/CompanySettingsTermsOfUseFormNavigatorEntry.java
@@ -12,16 +12,25 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.portalsettings.web.servlet.taglib.ui;
 
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
+
+import javax.servlet.ServletContext;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Pei-Jung Lan
+ * @author Eduardo Garcia
+ * @author Philip Jones
  */
-@OSGiBeanProperties(property = {"service.ranking:Integer=40"})
-public class CompanySettingsMailHostNamesFormNavigatorEntry
+@Component(
+		immediate = true, property = {"service.ranking:Integer=50"},
+		service = FormNavigatorEntry.class
+)
+public class CompanySettingsTermsOfUseFormNavigatorEntry
 	extends BaseCompanySettingsFormNavigatorEntry {
 
 	@Override
@@ -32,12 +41,21 @@ public class CompanySettingsMailHostNamesFormNavigatorEntry
 
 	@Override
 	public String getKey() {
-		return "mail-host-names";
+		return "terms-of-use";
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.portal.settings.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
 	}
 
 	@Override
 	protected String getJspPath() {
-		return "/html/portlet/portal_settings/mail_host_names.jsp";
+		return "/terms_of_use.jsp";
 	}
 
 }
