--- portal-impl/src/META-INF/form-navigator-spring.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-
-<beans
-	default-destroy-method="destroy"
-	default-init-method="afterPropertiesSet"
-	xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
->
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsAdditionalEmailAddressesFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsAdditionalEmailAddressesFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsAddressesFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsAddressesFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsAnalyticsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsAnalyticsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsAuthenticationFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsAuthenticationFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsConfigurationFormNavigatorCategory" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsConfigurationFormNavigatorCategory" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsContentSharingFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsContentSharingFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsDisplaySettingsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsDisplaySettingsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsEmailNotificationsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsEmailNotificationsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsGeneralFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsGeneralFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsIdentificationFormNavigatorCategory" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsIdentificationFormNavigatorCategory" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsInteractionsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsInteractionsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsMailHostNamesFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsMailHostNamesFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsMapsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsMapsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsMiscellaneousFormNavigatorCategory" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsMiscellaneousFormNavigatorCategory" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsPhoneNumbersFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsPhoneNumbersFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsRatingsFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsRatingsFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsRecycleBinFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsRecycleBinFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsSocialFormNavigatorCategory" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsSocialFormNavigatorCategory" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsTermsOfUseFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsTermsOfUseFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsUsersFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsUsersFormNavigatorEntry" />
-	<bean id="com.liferay.portal.servlet.taglib.ui.CompanySettingsWebsitesFormNavigatorEntry" class="com.liferay.portal.servlet.taglib.ui.CompanySettingsWebsitesFormNavigatorEntry" />
-</beans>
