+++ portal-impl/src/portal.properties
@@ -836,7 +836,6 @@
         META-INF/cluster-spring.xml,\
         META-INF/comment-spring.xml,\
         META-INF/document-library-spring.xml,\
-        META-INF/form-navigator-spring.xml,\
         META-INF/layout-spring.xml,\
         META-INF/message-boards-spring.xml,\
         META-INF/model-listener-spring.xml,\
